+++ packages/gov-ui-kit/src/core/components/dialogs/dialog/dialogHeader/dialogHeader.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from '@testing-library/react';
+import { render, screen, within } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
 import { IconType } from '../../../icon';
 import { DialogRoot } from '../dialogRoot';
@@ -6,55 +6,42 @@ import { DialogHeader, type IDialogHeaderProps } from './dialogHeader';
 
 describe('<Dialog.Header/> component', () => {
     const createTestComponent = (props?: Partial<IDialogHeaderProps>) => {
-        const completeProps: IDialogHeaderProps = { title: 'title', description: 'test', ...props };
+        const completeProps: IDialogHeaderProps = {
+            title: 'title',
+            ...props,
+        };
 
         return (
-            <DialogRoot open={true}>
-                <DialogHeader {...completeProps} />;
+            <DialogRoot hiddenDescription="description" open={true}>
+                <DialogHeader {...completeProps} />
             </DialogRoot>
         );
     };
 
-    it('renders the given title and description', () => {
+    it('renders the given title', () => {
         const title = 'test title';
-        const description = 'test description';
-
-        render(createTestComponent({ title, description }));
-
+        render(createTestComponent({ title }));
         expect(screen.getByText(title)).toBeInTheDocument();
-        expect(screen.getByText(description)).toBeInTheDocument();
-
-        const dialog = screen.getByRole('dialog');
-        expect(dialog).toHaveAccessibleName(title);
-        expect(dialog).toHaveAccessibleDescription(description);
+        expect(screen.getByRole('dialog')).toHaveAccessibleName(title);
     });
 
-    it('renders a back button when showBackButton is set to true', () => {
-        render(createTestComponent({ showBackButton: true }));
-
-        const backIcon = screen.getByTestId(IconType.CHEVRON_LEFT);
-        // eslint-disable-next-line testing-library/no-node-access
-        expect(backIcon.closest('button')).toBeInTheDocument();
+    it('does not render a close button when the onClose property is not set', () => {
+        render(createTestComponent());
+        expect(screen.queryByRole('button')).not.toBeInTheDocument();
     });
 
-    it('calls onBackClick when the back button is clicked', async () => {
-        const user = userEvent.setup();
-        const handleBackClick = jest.fn();
-
-        render(createTestComponent({ showBackButton: true, onBackClick: handleBackClick }));
-
-        const backIcon = screen.getByTestId(IconType.CHEVRON_LEFT);
-        await user.click(backIcon);
-
-        expect(handleBackClick).toHaveBeenCalled();
+    it('renders a close button when the onClose property is set', () => {
+        const onClose = jest.fn();
+        render(createTestComponent({ onClose }));
+        const closeButton = screen.getByRole('button');
+        expect(closeButton).toBeInTheDocument();
+        expect(within(closeButton).getByTestId(IconType.CLOSE)).toBeInTheDocument();
     });
 
-    it('calls onCloseClick when the close button is clicked', async () => {
-        const handleOnCloseClick = jest.fn();
-
-        render(createTestComponent({ onCloseClick: handleOnCloseClick }));
-
+    it('calls onClose when the close button is clicked', async () => {
+        const onClose = jest.fn();
+        render(createTestComponent({ onClose }));
         await userEvent.click(screen.getByRole('button'));
-        expect(handleOnCloseClick).toHaveBeenCalled();
+        expect(onClose).toHaveBeenCalled();
     });
 });