+++ packages/gov-ui-kit/src/core/components/dialogs/dialog/dialogHeader/dialogHeader.tsx
@@ -1,69 +1,52 @@
-import { Close, Description, Title } from '@radix-ui/react-dialog';
+import { Close, Title } from '@radix-ui/react-dialog';
+import classNames from 'classnames';
 import { type ComponentPropsWithoutRef } from 'react';
-import { Button } from '../../../button';
+import { AvatarIcon } from '../../../avatars';
 import { IconType } from '../../../icon';
 
 export interface IDialogHeaderProps extends ComponentPropsWithoutRef<'div'> {
     /**
-     * Optional accessible description announced when the dialog is opened
-     */
-    description?: string;
-    /**
-     * Indicates whether a back button should be shown
-     * @default false
-     */
-    showBackButton?: boolean;
-    /**
-     * Accessible title summarizing dialog's content or purpose. Will be announced when
-     * dialog is opened.
+     * Title of the dialog displayed on the header and used as the dialog's accessible name.
      */
     title: string;
     /**
-     * Callback invoked when the back button is clicked
+     * Callback triggered on close button click. The close button is not displayed when the property is not set.
      */
-    onBackClick?: () => void;
-    /**
-     * Callback invoked when the close button is clicked. Closes the dialog by default
-     */
-    onCloseClick?: () => void;
+    onClose?: () => void;
 }
 
-/**
- * `Dialog.Header` component
- *
- * **NOTE**: This component must be used inside a `<Dialog.Root />` component.
- */
 export const DialogHeader: React.FC<IDialogHeaderProps> = (props) => {
-    const { description, showBackButton = false, title, onBackClick, onCloseClick, ...otherProps } = props;
+    const { title, onClose, className, ...otherProps } = props;
+
+    const headerClassNames = classNames(
+        'relative flex w-full items-start rounded-t-xl bg-modal-header backdrop-blur-md', // Layout
+        'pb-1.5 pl-4 pr-14 pt-4 md:pb-2 md:pl-6 md:pr-16 md:pt-6', // Spacings
+        className,
+    );
+
+    const closeButtonClassNames = classNames(
+        'group rounded-full border border-neutral-100 bg-neutral-0 p-1 outline-none', // Default
+        'absolute right-3 top-3 md:right-4 md:top-4', // Positioning
+        'hover:border-neutral-200 active:border-neutral-200 active:bg-neutral-50', // Hover/Active states
+        'focus:outline-none focus-visible:ring focus-visible:ring-primary focus-visible:ring-offset', // Focus state,
+    );
 
     return (
-        <div
-            className="flex w-full items-start gap-x-3 rounded-t-xl bg-modal-header px-4 pb-2 pt-4 backdrop-blur-md md:gap-x-4 md:px-6 md:pt-6"
-            {...otherProps}
-        >
-            {showBackButton && (
-                <Button
-                    variant="tertiary"
-                    size="sm"
-                    iconLeft={IconType.CHEVRON_LEFT}
-                    className="shrink-0"
-                    onClick={onBackClick}
-                />
-            )}
-            <div className="min-w-0 flex-1 gap-0.5">
-                <Title className="flex-1 truncate text-lg leading-tight text-neutral-800">{title}</Title>
-                {description && (
-                    <Description className="flex-1 text-sm leading-normal text-neutral-500">{description}</Description>
-                )}
-            </div>
+        <div className={headerClassNames} {...otherProps}>
+            <Title className="flex-1 truncate text-lg font-normal leading-tight text-neutral-800 md:text-xl">
+                {title}
+            </Title>
             <Close asChild={true}>
-                <Button
-                    variant="tertiary"
-                    size="sm"
-                    iconLeft={IconType.CLOSE}
-                    className="shrink-0"
-                    onClick={onCloseClick}
-                />
+                {onClose != null && (
+                    <button onClick={onClose} className={closeButtonClassNames} type="button">
+                        <AvatarIcon
+                            icon={IconType.CLOSE}
+                            size="sm"
+                            backgroundWhite={true}
+                            className="group-hover:bg-neutral-50"
+                        />
+                    </button>
+                )}
             </Close>
         </div>
     );