+++ packages/gov-ui-kit/src/core/components/dialogs/dialog/dialogRoot/dialogRoot.test.tsx
@@ -1,35 +1,26 @@
 import { render, screen } from '@testing-library/react';
-import { DialogHeader } from '../dialogHeader';
 import { DialogRoot } from './dialogRoot';
 import type { IDialogRootProps } from './dialogRoot.api';
 
 describe('<Dialog.Root/> component', () => {
-    const createTestComponent = (rootProps?: Partial<IDialogRootProps>) => {
-        const completeRootProps: IDialogRootProps = {
-            ...rootProps,
+    const createTestComponent = (props?: Partial<IDialogRootProps>) => {
+        const completeProps: IDialogRootProps = {
+            ...props,
         };
 
-        return <DialogRoot {...completeRootProps} />;
+        return <DialogRoot hiddenTitle="title" hiddenDescription="description" {...completeProps} />;
     };
 
     it('does not render the dialog by default', () => {
         render(createTestComponent());
-
         expect(screen.queryByRole('dialog')).not.toBeInTheDocument();
     });
 
     it('renders the dialog with the given content', () => {
-        const content = (
-            <>
-                <DialogHeader title="test" description="description" />
-                <div>test content</div>
-            </>
-        );
-
-        render(createTestComponent({ open: true, children: content }));
-
+        const children = 'test-content';
+        render(createTestComponent({ open: true, children }));
         const dialog = screen.getByRole('dialog');
         expect(dialog).toBeInTheDocument();
-        expect(screen.getByText('test content')).toBeInTheDocument();
+        expect(screen.getByText(children)).toBeInTheDocument();
     });
 });