+++ packages/gov-ui-kit/src/core/components/dialogs/dialogAlert/dialogAlertFooter/dialogAlertFooter.test.tsx
@@ -1,5 +1,4 @@
 import { render, screen } from '@testing-library/react';
-import { DialogAlertHeader } from '../dialogAlertHeader';
 import { DialogAlertRoot } from '../dialogAlertRoot';
 import { DialogAlertFooter, type IDialogAlertFooterProps } from './dialogAlertFooter';
 
@@ -12,8 +11,7 @@ describe('<DialogAlert.Footer/> component', () => {
         };
 
         return (
-            <DialogAlertRoot open={true}>
-                <DialogAlertHeader title="title" description="description" />
+            <DialogAlertRoot hiddenTitle="title" hiddenDescription="description" open={true}>
                 <DialogAlertFooter {...completeProps} />;
             </DialogAlertRoot>
         );