+++ packages/gov-ui-kit/src/core/components/dialogs/dialogAlert/dialogAlertFooter/dialogAlertFooter.tsx
@@ -1,63 +1,53 @@
 import { Action, Cancel } from '@radix-ui/react-alert-dialog';
 import classNames from 'classnames';
 import type React from 'react';
-import { useContext, type AnchorHTMLAttributes, type ButtonHTMLAttributes } from 'react';
-import { Button, type IButtonBaseProps } from '../../../button';
-import { DialogAlertContext } from '../dialogAlertRoot';
+import { Button, type IButtonProps } from '../../../button';
+import { useDialogAlertContext } from '../dialogAlertContext';
 
-export type IDialogAlertFooterAction = (
-    | Pick<AnchorHTMLAttributes<HTMLAnchorElement>, 'href'>
-    | Pick<ButtonHTMLAttributes<HTMLButtonElement>, 'onClick'>
-) &
-    Pick<IButtonBaseProps, 'iconRight' | 'iconLeft' | 'disabled' | 'isLoading'> & {
-        /**
-         * Button label
-         */
-        label: string;
-    };
+export type IDialogAlertFooterAction = Exclude<IButtonProps, 'children' | 'variant'> & {
+    /**
+     * Label of the action button.
+     */
+    label: string;
+};
 
 export interface IDialogAlertFooterProps {
     /**
-     * Alert dialog primary action button
+     * Action button of the alert dialog.
      */
     actionButton: IDialogAlertFooterAction;
     /**
-     * Alert dialog secondary button used for dismissing the dialog
-     * or cancelling the action
+     * Secondary button of the alert dialog used for dismissing the dialog or cancelling the action.
      */
     cancelButton: IDialogAlertFooterAction;
 }
 
-/**
- * `DialogAlert.Footer` component
- *
- * **NOTE**: This component must be used inside a `<DialogAlert.Root />` component.
- */
 export const DialogAlertFooter: React.FC<IDialogAlertFooterProps> = (props) => {
-    const { variant } = useContext(DialogAlertContext);
-
     const { actionButton, cancelButton } = props;
+
+    const { variant } = useDialogAlertContext();
+
     const { label: actionLabel, ...actionButtonProps } = actionButton;
     const { label: cancelLabel, ...cancelButtonProps } = cancelButton;
 
     const actionVariant = variant === 'info' ? 'secondary' : variant;
+    const reverseButtonOrder = variant === 'warning' || variant === 'critical';
+
+    const footerClassNames = classNames(
+        'flex gap-3 rounded-b-xl bg-modal-footer px-4 pb-4 pt-3 backdrop-blur-md md:gap-4 md:px-6 md:pb-6',
+        { 'flex-col md:flex-row': !reverseButtonOrder },
+        { 'justify-end flex-col-reverse md:flex-row-reverse': reverseButtonOrder },
+    );
 
     return (
-        <div className="flex flex-col gap-3 rounded-b-xl bg-modal-footer px-4 pb-4 pt-2 backdrop-blur-md md:flex-row md:px-6 md:pb-6">
+        <div className={footerClassNames}>
             <Action asChild={true}>
-                <Button
-                    {...actionButtonProps}
-                    size="lg"
-                    variant={actionVariant}
-                    className={classNames('w-full md:w-auto', {
-                        'order-2': variant === 'warning' || variant === 'critical',
-                    })}
-                >
+                <Button size="md" {...actionButtonProps} variant={actionVariant}>
                     {actionLabel}
                 </Button>
             </Action>
             <Cancel asChild={true}>
-                <Button {...cancelButtonProps} variant="tertiary" size="lg" className="w-full md:w-auto">
+                <Button size="md" {...cancelButtonProps} variant="tertiary">
                     {cancelLabel}
                 </Button>
             </Cancel>